Update web app Dockerfile
This brings the file more inline with what I did in [simple-sinatra-docker](https://github.com/DEFRA/simple-sinatra-docker) which took as its base the example given in [Ruby docker hub entry](https://hub.docker.com/_/ruby).

It also aims to follow what seems to be the general advice, that of ordering the actions in your Dockerfile by need to change.