First steps to a production build
So I'm aiming to understand how to put together a production build for one of our services.

To that end I've split the files and updated the docker-compose.yml. I'm aware that in the end I should come back to a multi-stage build to handle dev and prod environments. But for now I just want to keep things simple and focused.