Update and amend gitignore
In preparation for supporting running the web app locally against the docker-compose PostgreSQL container (allows me to debug issues more easily) this makes some changes to our gitignore file.

First of all we expand the list of things we want to ignore based on experience in other projects.

Second we start allowing all our .env files to be committed. This would be a big no-no in an actual repo. But we're just testing and exploring so this allows me to see everything in the project.

Lastly we add a .env.development.local, which will take precedence when we incorporate the [Dotenv](https://github.com/bkeepers/dotenv#what-other-env-files-can-i-use) gem.