Add Dotenv gem
This is what will allow us to run the app locally more easily when needed.

As dotenv is only in the development and test groups, the gem will not get loaded in the production version of the container we're building so we can still be confident values are being passed into the container as environment variables.

We intend to add a .dockerignore file soon to ensure .env files don't get copied to the container to confirm this.