Forward localhost 54320 to postgres container
To enable us to talk to the postgres container docker-compose spins up we need to tell docker-compose to map a host port to the port in the container.

Via docker magic we don't have to worry about telling other containers in our compose file about this. As they are on the same network they can talk directly to 5432.

This is purely about enabling connection from the host to the container.

N.B. expressing it as a string is a recommendation from Docker https://docs.docker.com/compose/compose-file/#ports