Implement puma as application server
If we were deploying into a normal production environment we would be expected to implement some form of application server as webrick is not suitable.

We have used passenger in the past but it has always been out intention to move to something simplier to manage, and full open source. Hence we go with puma (which is ther default app server in later versions of Rails).

TBH there wasn't a very clear and clean example to use for the config when the intent is to host the app in docker. But based on our 'brief' research what we have done here is appropriate and suitable for the kind of app we are running.