Switch to using volumes in docker-compose
Previously for our database we had created a mount point between the container and the host system. Though we could be relative, it still depends a certain knowledge of the host system.

Where data needs to be shared between containers Docker recommends using volumes instead

> Volumes are the preferred mechanism for persisting data generated by and used by Docker containers.

https://docs.docker.com/storage/volumes/

Essentially you ask docker to create some 'magic' space on the host which we can reference simply using a name. This then makes it easily portable between environments as you need to know nothing in order to have somewhere to save generated data and make it available between containers.

This change switchs the postgres container to use a volume. We also create a `public_volume` with an eye to sharing our compiled assets between the app container and the nginx web container.

Oh, and that explains why we have also renamed the rails container!