Add nginx to environment (almost!)
So this adds nginx to the environment as the reverse proxy server. We use Nginx in our current production environments and using Nginx seems to be a standard pattern even when using Docker.

So with this in place and running, the test page is shown if you navigate to http://localhost (no need for a port numeber!) However no assets are served. To ensure we are getting assets from Nginx instead of the app itself we have removed the env var which tells Rails to serve static assets. And having manually copied the pre-compiled assets from the app server to the web server using `docker cp` I can confirm that once there Nginx will serve them correctly.

I just need to figure out how to get the app generating the assets at build and having them available when the web image is built.