Reduce volume exposure / best practice bash
Though we only really care about the `public/` folder everything from the repo seemed to be finding its way onto the web image. So I made this change to make what we actually need more explicit and it seemed to do the trick!

Also made a quick tweak to the entrypoint.sh script. It seems best practice is to start all bash scripts with `set -euo pipefail`. To explain

- `e` exit on first error
- `u` exit when an undefined variable such as $FO0 is accessed
- `o pipefail` exit when | any |cmd | in | a | pipe has exitcode != 0

So this is a way of ensuring that everything in the script works and that we exit as soon as anything fails.