Add healthcheck to the NGINX docker file
Came across the mention of 'wget healthchecks' here https://nickjanetakis.com/blog/best-practices-when-it-comes-to-writing-docker-related-files. Then tracked the reference down https://docs.docker.com/engine/reference/builder/#healthcheck

Figured it couldn't hurt to have one so added it to the Dockerfile.

N.B. Also found an alternative version here https://medium.com/@ryanblunden/for-those-looking-for-a-http-health-check-in-alpine-images-wget-is-available-from-the-base-image-a07997be8699

```
HEALTHCHECK --interval=5s --timeout=5s --retries=3 CMD wget localhost:8080/healthz -q -O - > /dev/null 2>&1
```

When I ran the wget command when `docker exec` onto the container this worked, but didn't seem to cause any visible response in the docker-compose log output. But when I applied it to the Dockerfile it did work. So it might be better. But for now sticking with a version I can understand and see doing something when I run it directly.