Add a healthcheck for the app container
The healthcheck in the app was cribbed from https://docs.publishing.service.gov.uk/manual/setting-up-new-rails-app.html and gives us something light weight that doesn't reference any other dependencies or assets.

We then just apply the same check we implemented for the Nginx container.

N.B. Also cleaned out all the comments as they are not relevant to what we are doing.