Slim down rails app image
These changes focus on reducing the size of the image our current rails app Dockerfile creates.

The main change was switching from the standard debian image to one based on Alpine. This is a different OS hence commands like apt-get needed to be updated, and to keep its size small tools like bash are not installed.

Along the way we improved out understanding of the intermediate image layers created when building the final image, and so have tweaked the file to ensure these are kept to a minimum.

The issue of there being no bash is also what tipped out hand to remove the ENTRYPOINT. In our production container we should never encounter any other pids plus this is not a strategy I have seen across the examples I have looked at. And we bash no longer being available it would mean a re-write. So instead we have just dropped it.

For reference we have gone from an image size of 839MB to one that is 239MB!