Switch to multistage build
https://docs.docker.com/develop/develop-images/multistage-build/
https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#use-multi-stage-builds
https://thecode.pub/reduce-your-docker-images-an-example-with-ruby-564735f4388c

In my original work with Docker each environment e.g. dev, test, production etc had their own Dockerfile. This led to duplication and bloated images so the builder pattern was formed. This essentially breaks down the steps of a build into separate files, and would then use bash scripts to 'glue' them together to generate your final dev, test or production image.

Then a couple of years ago Docker released multistage builds. This allowed you to do the same thing, but within a single DOckerfile, with proper support for linking the different images together. Less chance of errors and no need for bash scripts anymore.

It also improves the use of cached layers when having to generate builds, and often results in an even smaller final image. In our case we've got the rails app image down to 172MB for production with this change.

So this change updates the project to use a multistage Dockerfile, as per current best practise.