Amend ports to support scaling
So the `ports` declaration in the docker-compose.yml is used to hook up the host to the docker container and allow us to access it. Hence for a rails app running in a container we would expect to see `"3000:3000"` (host port:container port).

That's fine for a service we only expect to run one instance of. But one of the benefits of Docker and Docker compose is it allows us to easily run multiple instances of a service. If we are doing that, there is no way we can talk to them all on port `3000`.

So compose has this awesome ability to be able to dynamically assign host ports for each instance started. So if instead we set ports to `"3000"` docker compose will assign its own host port for each instance started that connects to 3000 on each of them.

You can find the host ports (so you can still connect directly if needed) by using either `docker-compose ps` or `docker ps`.

For reference, to run multiple instances of our app service we would do `docker-compose up --scale app=5`.