Remove healthchecks
This is a bit of an arse, but I have spotted since switching to alpine as the base image the healthchecks no longer work. If I `docker exec` onto the instance I can replicate the wget calls fine. But the healthchecks report the same thing in both

```json
{
  "Start": "2020-02-27T13:02:17.0507153Z",
  "End": "2020-02-27T13:02:17.2414162Z",
  "ExitCode": 1,
  "Output": "wget: bad address '|| exit 1'\n"
},
```

I believe the problem is a known issue with the DNS in Alpine images. What I'm struggling with right now is trying to find a clean example of what I need to do to resolve it

- https://medium.com/@xavier.priour/docker-alpine-dns-issue-bad-address-84594d128d9f
- https://forums.docker.com/t/resolved-service-name-resolution-broken-on-alpine-and-docker-1-11-1-cs1/19307
- https://github.com/gliderlabs/docker-alpine/issues/476
- https://unix.stackexchange.com/questions/441664/alpine-linux-sometimes-dns-is-not-resolved
- https://github.com/docker/for-linux/issues/755
- https://stackoverflow.com/questions/57202039/resolve-conf-cant-be-changed-docker-alpine
-

I have also tried playing with and removing the rails user (in case it was a permissions issue) and carrying out a `apk upgrade -U -a` as part of the build to ensure everything in the image is the latest and greatest but still no joy.

So as I never actually see these when I'm googling for examples, and I know the apps are currently working, I'm removing them for now. I would like to bring them back in later though once I've got a bit more time to look into the problem.