Move nginx into multi-stage build
This moves the build instructions for the Nginx image into our main Dockerfile. Doing so makes it easier to fit it into our multistage build. So now rather than needing to share a volume between it and the ruby image to get the assets, it can copy them out of the rails_builder stage directly.