Make changes to reduce when re-builds happen
Because we switched to copying across the whole root of the current project as part of the rails_builder stage I spotted this meant we were invalidating our image and causing it to build any time we were making a change to the code base.

So to reduce the chance of Docker deciding that we need to rebuild this stage, this change only copies across the bear minimum needed to install the gems and generate the assets.

From here it should mean you can code away, and in most cases the only thing that will need to be updated is the production stage, which by this point is doing very little.