Add mailcatcher to environment plus stage updates
We use mailcatcher in development as a means of testing and seeing the emails our apps send. This means creating an image so we can run   it in a container.

In order to do this though, I needed to install the mailcatcher gem, which includes dependencies on ones with native extensions. So i     needed the same build dependencies used to install the gems for the project.

Rather than duplicate what I was already doing in that image, I chose to break the stages down further and reuse them. So a lot of the changes here are about refactoring the stages to improve reuse.