Use mailcatcher image and revert Dockerfile
Realised I was being a donkey. Why should I maintain and create a mailcatcher when surely someone has aleady done that!

Turns out they have. So this updates the docker-compose to use the image. Plus we revert our Dockerfile. With no need for such broken down stages, and little likelihood we will, there is no sense in having stages around 'just in case'.