Move mailcatcher up in docker-compose
Following more best practise, specifically to order the services in your docker-compose file in the order you need them to be started.

It's not critical in our case that mailcatcher is second. But ideally it should be available from the moment the app service is up and runniing, so moving it above the app service ensures it's at least started first.

https://nickjanetakis.com/blog/best-practices-when-it-comes-to-writing-docker-related-files
https://docs.docker.com/compose/compose-file/#depends_on