Keep lib/assets in production image
Actually came up whilst working on a working email example. I'm not sure why it was suggested that `lib/assets` be deleted in a running production app when after a google I found

> `lib/assets` is for your own libraries' code that doesn't really fit into the scope of the application or those libraries which are shared across applications.
https://guides.rubyonrails.org/asset_pipeline.html#asset-organization

The examples it provides focus on javascript files, but from some more quick googling the `lib` is for any content that does not relate directly to your app, and that does not come from an external vendor. So there is a good chance this is actually something you'll want to keep on the production iimage. Plus in my upcoming changes for email, I needed somewhere to store a png file which would be a test attachment. We couldn't use `app/assets` as that is deleted on the basis the assets have been pre-compiled and added to the web service image. Also `public/` is deleted because again it is loaded into the web serice image. So `lib/assets` was the best option.