Implement mail sending example
The following change allows me to test sending an email from the running app container to the mail service (what would be an external service like Sendgrid in real life). This is the same setup we use when running locally as well so its a handy way of ensuring we have a working example of that.

For now to send the email requires `docker exec -it simple-rails-docker_app_1 bundle exec rake email:test` but its a start.

N.B. Along with the previous changes email sending is working and being received by the mail service, accessible at http://localhost:1080