Allow port to be configurable in puma config
Whilst working on an admin container for running cron jobs and one off tasks we originally thought we would have to start rails in order to have a long live process (if you don't call one in the CMD the container will immediately close). But we wouldn't want it responding to requests from Nginx so we had a need to assign a different port.

That need is no longer there, but being able to configure the port has value so we are leaving this change as is.