Complete working admin container with cronjob
The goal of this and the last few builds was to get to a point where we had a container which when run would be used for admin tasks, such as `bundle exec rake db:migrate`. In our environment it would also be where scheduled tasks would be kicked off.

We wanted to retain the use of the 'whenever' gem as its great to be able to use ruby to define the cron schedule. Also as part of our normal deployment (we use Capistrano) it would handle creating or updating the crontab for our app.

This is what we set out to replicate in our container; using whenever to read our schedule and update the crontab file, and cron to kick off our rake tasks when scheduled.

However, though there were examples of creating containers with cron running jobs, generally advice is to avoid doing it. The ideal is to have something outside the Docker eco-system that is handling the schedule, and when called fires up your container with the specific rake command.

As we don't know what that could be, and it would mean giving up control of the scheduling and running of the jobs (in our specific case) we ignored this.

The next issues are that cron generally expects to be running as the root user, or at least one with admin privileges. Running as root is not best practise for docker containers, so we couldn;t do this. There appears to be ways to have cron run specific jobs as different users without admin permissions. But then we encountered other issues with using cron, succinctly listed here

- They purge their environment before starting jobs. This is an important security feature in multi-user systems, but it breaks a fundamental configuration mechanism for containers.
- They capture the output from the jobs they run, and often either want to email this output or simply discard it. In a containerized environment, logging task output and errors to `stdout` / `stderr` is often easier to work with.
- They often don't respond gracefully to `SIGINT` / `SIGTERM`, and may leave running jobs orphaned when signaled. Again, this makes sense in a server environment where `init` will handle the orphan jobs and Cron isn't restarted often anyway, but it's inappropriate in a container environment as it'll result in jobs being forcefully terminated (i.e. `SIGKILL`'ed) when the container exits.
- They often try to send their logs to syslog. This conveniently provides centralized logging when a syslog server is running, but with containers, simply logging to `stdout` or `stderr` is preferred.

This list comes from [Supercronic](https://github.com/aptible/supercronic) which was referenced in a few of the examples we found.

Quoting Supercronic

> Supercronic's goal is to behave exactly how you would expect cron running in a container to behave:

> - Your environment variables are available in jobs
> - Job output is logged to `stdout` / `stderr`
> - `SIGTERM` triggers a graceful shutdown (and so does `SIGINT`, which you can deliver via CTRL+C when used interactively)
> - Job return codes and schedules are logged to `stdout` / `stderr`
> - `SIGUSR2` triggers a graceful shutdown and reloads the crontab configuration

So the final result is we build a container that still uses whenever to generate the crontab file, but uses supercronic to trigger the jobs according to the schedule.

To cater for this we also amended how we write our `schedule.rb`. Behind the scenes whenever is normally doing clever things to ensure your command runs with the same env vars available as your app. As this is all handling for us we need to tell whenever to just run our command as stated.

That's about it. With the setup currently a `docker-compose build && docker-compose up` should result in the Nginx reverse-proxy running, our app, PostgreSQL server for the database, Mailcatcher for testing emails, and the admin container with our app loaded, but running supercronic to trigger a test email every minute.