Add addressbase Java container to environment
This change adds a new java based container to our Dockerfile which will be used for running the Addressbase Facade jar file.

Our main issue with this is the ABF's reliance on config files rather than environment variables to drive configuration, specifically getting the OS Places API key to be used.

We can't put the key in the file directly, as we need to store `config.json` in source control. We could have used steps to modify the file as part of the image build, and then we only have to deal with the issue once (unless the key changed). However this would mean the image is compromised because it includes a copy of our OS Places key.

So instead we have had to resort to a solution where the `config.json` file is generated fresh each time an image is executed as a container. In this way we can stick to the convention that the OS Places API key need only be passed in at run time.