Start ignoring .env file again
We are on the cusp of updating the repo to allow multiple different environments to be started using a feature ion docker-compose that allows mulitple compose files to be used in combination.

As such this change moves the .env file back to being the default file for local development. We'll create/update the rest to match the environments we'll create using this new mechanism and leave those as part of the source code.