Switch to using multiple docker-compose files
Docker-compose has a natty feature which allows you to specify a root docker-compose file, and then multiple override files <https://docs.docker.com/compose/reference/overview/#specifying-multiple-compose-files>

This is awesome, as it allows us to support building and running our app in multiple environments. For example in our project we always need a database and an address lookup. But we only care about the complexity of a load balancer when running in "production" mode. Else we want the app running in "development" mode so we can access it directly, and see things like errors displayed in full.

We can do this by using commands like `docker-compose -f docker-compose.yml -f docker-compose.production.yml up`.

So this change includes splitting the docker-compose.yml file into 3. We also made some changes to the Dockerfile. Specifically we have added a new stage which installs packages needed to install our gems. This can then be used directly by the development build, or indirectly via the builder stage when creating the production build.

It also includes separate .env files for each environment (though at this time the content is the same).

Note. Not everything will work as expected. We need to make some changes to the app the Docker files to truly allow flexibility in running multiple environment, including supporting running the app locally but still using resources like the database and the address lookup from docker.

Plus the commands are now getting quite long and complex just to start an environment. So the next steps are to add support for those things.