Add helper script for working with Compose
We are currently driving everything through compose because we are thinking in terms of environments (local, development, production etc) rather than just individual containers.

But to support these multiple environments we need to leverage some features of Compose that in turn greatly lengthen and complicate the commands we have to run. Also we know there will be team members new to Docker and Compose so having a script that helps simplify the commands needed to start working with the project will be a great help.

Further details about the features of Compose we are using are detailed in the comments of the doc.sh file.

NOTE. Work still needs to be done to all the app work across the different environments by making the env vars and the way it uses them more flexible.