Ensure webrick is used in development
There seems to be a known issue that even if puma sits within a :production group in  your Gemfile it will still be used instead of webrick if the gem is installed. Running with Puma is probably fine, but when working in development our expectation is that webrick will be used, and the app will behave as such. So to stick with this convention, we ensure any production only gems are not installed in the development build.

N.B. This has nothing to do with docker, other than we switched to Puma as part of this experiment and this is the first time we have noticed this!