Update envs correctly for each environment
This preps the .env files ready for getting all 3 environments working

- local
- development
- production

The key correction we needed to make was due to our lack of understanding about how .env prioritises files when running locally.

We were mistakenly setting `.env` to be the file used when running locally, and the `.env.development` and `.env.production` ones just to be referenced by Compose. However the actual order of priority is

- 1st - `.env.development.local` (highest)
- 1st - `.env.test.local`
- 1st - `.env.production.local`
- 2nd - `.env.local`
- 3rd - `.env.development`
- 3rd - `.env.test`
- 3rd - `.env.production`
- last - `.env`

Plus a bit like Compose's merging of multiple files, dotenv merges the values in these files. So a value set in .env will be overridden by one set in a file higher up the ordering.

This is exactly what was happening in our case. Though we had configured `.env` to have values relevant to running the app locally on the host, these were being overridden by the values in `.env.development`.

So going foward, because we are trying to get this to work both locally and in Docker, we'll be ignoring completely the `.env` file. Instead when running locally we'll create a `.env.local` file. We'll keep the `.env.[environment]` files setup for their relevant Docker environments.