Make mail service available to the host
We had forwarded port 1080 to the host so we could access the web UI for mailcatcher. However we hadn't forwarded 1025. This meant when running the app locally it couldn't connect to send emails.

Also we have updated both declarations to follow best practise and wrap them in quotes to avoid any Yaml funnies with numbers.