Update app settings 4 env vars / remove defaults
This is the final change needed to get all 3 environments working (local, development and production).

The key change was using an env var for the address lookup url. But to ensure we are seeing the env vars and not just getting by because we have a default, this also removes all env var defaults.