Change cron schedule to every 5 minutes
Only because it was creating too much noise when running the production environment. Seeing it happen every 5 minutes is enough to prove its working.

N.B. Whenever will generate a cron to run the job at every 5 minutes passed the hour, not every 5 minutes from the time the environment is started.