Bind mount project rather than copy for development
In the development build we want to provide an environment where development can take place but the user/developer does not have to install anything locally if that is the workflow they choose to use.

Instead if you were to run this container you would be expected to bind mount a local folder to the container when it ran https://docs.docker.com/storage/bind-mounts/

Because we intend to use Compose in most cases, this project's development docker-compose.yml (docker-compose.development.yml) handles setting up the bind mount.

This means when the container is run, the current directory is shared with the running container, and means we can do things like editing files locally and see the changes in the running instance in the development container.