Add helper command to access a running container
Often for debug purposes and just to check out what is going on its handy to access a running container and have a poke around.

The Compose command is essentially `exec my_container /bin/sh` and we would support this using the existing `exec` action built into our doc.sh.

But for new users who might not be aware of how to do this, and to document how you'd go about doing it (plus handling things like when a service is scaled) we have added the helper action `dive` to our script.