Switch from Minitest to rspec
Did actually want to give minitest a go, but after adding a test nothing seemed to be happening when I ran `bundle exec rake minitest`.

I didn't want to waste too much time on this, and with all our apps using rspec it actually made sense to switch things over so this project more closely follows them.

So as a first step to using Docker for our unit tests, this updates to the app to use rspec instead of minitest, and adds a basic request test to get things rolling.