Make logging to STDOUT configurable
We hit a snag now tests are running, especially request tests.

Because the logger had been set to STDOUT at the application level our rspec output was being polluted by everything rails would normally output to `log/test.log`.

So this highlighted that we need to make logging to STDOUT configurable in the app. This change adds a new env var that all the Docker builds set, but which we wouldn't expect to set locally.

We didn't put this in the environment .env files because its not something we expect anyone to want to configure (logging to STDOUT is best practise and expected for apps running in containers). That's why its set in the Dockerfile.

We also could have set this once in `rails_base` in the Dockerfile. But when it comes to the test stage we intend to add, we wouldn't want that env var set. Instead we'll be looking to use something else to tell rspec to send its output to STDOUT so we can see it from outside the Docker container. Else we'd be back to what we have locally, a mix of rails output and rspec output all munged together.