Update project to support a test environment
This is the final step in supporting a test environment alongside development and production (Note. This breaks local but we're gonna fix that in the next commits!)

The need for a `.env.test` and a `.docker-compose.test.yml` are standard requirements by now. Only thing of note is the fact we don't bother to expose a port for the test app as the intent is to run our automated tests.

We have also updated the `doc.sh` helper script to support the new test environment. So it will now correctly select the docker-compose.yml file to use, and formate the Compose command for actions like `build`, `up` etc.

Key changes are in the Dockerfile, and the addition of an `entrypoint.sh`.

The `rails_test` app uses the default CMD of `bundle exec rspec`. The problem on first run (`up`) the app doesn;t stay around long enough for us to 'prep' the environment i.e. create the DB and run migrations.

We could get around this by manually overriding the CMD in `rails_test` to something long lived, then running our `./doc.sh test prep`, then restarting everything as normal. But just typing it out shows the process would be messy.

So we moved to just accepting that it'll be cleaner and easier if the development and test environments always ran `rake db:create` and `rake db:migrate` as part of their normal startup. This way they would just 'work' whether it was the first run or the 100th run.

Going forward, this should mean you'd only need to use `./doc.sh prod prep` for production.

The other key decision we made was to stop worrying about using Puma instead of Webrick. This significantly reduces the complexity of the Dockerfile and the `rails_development` and `rails_test` stages, and makes `rails_builder` much more useful and reusable.

What we just need to do going forward is make the Puma setup more configurable via env vars. So in dev and test we're just 1 thread max, but production we keep our current of 1 min and 5 max.

Testing shows it works the same, so we just need to get over ourselves and move on!