Move docker .env files to env folder
As noted in some if the previous commits, Dotenv and the .env files Compose uses were getting confused when run in certain modes. A file intended for Compose e.g. `.env.test` was being read by Dotenv when rspec was run locally.

Move these files which are intended only for Compose to a folder called env solves the problem, and makes their intent a little clearer.