Run nginx as non-root user
Again following best practices we want to ensure none of our containers are running as root. The last one to solve was NGINX.

These changes mean Nginx will now run as the `nginx` user, a non-root user that the official Nginx image provides.

Essentially the key blockers are

- you can no longer listen on port 80. You must pick a port greater than 1024
- the nginx user needs to have access and permissions to your web files, and certain files used when Nginx is running

The changes implemented here are taken from https://www.rockyourcode.com/run-docker-nginx-as-non-root-user/ with some modifications.

During this change also spotted a misunderstanding with the `ARG` command and value in the docker-compose file. I thought that Compose and docker was picking this up from the `env_file` we had declared. However it was actually picking it up from my .env file. This is why I have deleted that env var from all .env files and simply written it into the production docker-compose file. Hopefully we won't make that same mistake again.