Move and rename nginx.conf
As found when doing the research to get Nginx running as a non-root user, it heavily relies on 2 config files

- /etc/nginx/nginx.conf
- /etc/nginx/conf.d/default.conf

We have so far managed to not need to maintain our own version of nginx.conf and instead have focused on just having a default.conf template. This has everything we need to tell Nginx where our static assets are located, and what address to route all other requests through to.

However, we were calling our `default.conf` template `nginx.conf`! This is only likely to cause confusion. So this change creates a new nginx folder, so we its clear what the content is for. Plus we rename the file to match what it will be when copied to the docker image so its easier to see what its used for.