Add nginx.conf to solution
Having done an initial review of our Nginx setup and the one used in our WEX environments we have made these initial changes.

Key one is we are now doing other things in the `nginx.conf` so it makes sense to start maintaining our own version and copying it when we build our image. We have also tried to document it as best we can for future referencing.

The original WEX setup still has some differences, most noticeably it is setup for having Nginx load balance 2 apps, not just 1. But for now, we think this is sufficient until we start creating an actual Docker version of WEX.